# public/openapi-spec.yaml
openapi: 3.1.0
info:
  title: brodiegroch.ca API
  version: "1.0"
  description: Complete API for managing courses, deliverables, schedule, quick links, grading scale, and personal items
servers:
  - url: https://brodiegroch.ca
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
    Course:
      type: object
      properties:
        Course ID:
          type: string
          example: "COMP 213"
        Course Name:
          type: string
          example: "Computing for Engineering Technology"
        Course Average Weighted Grade:
          type: string
          example: "85.5"
        Credit Hours:
          type: string
          example: "3"
        Course Description:
          type: string
          example: "Course description here"
        Professor/Teacher Name:
          type: string
          example: "John Doe"
        Professor/Teacher Email:
          type: string
          format: email
          example: "john.doe@sait.ca"
...
    DetailedCourse:
      type: object
      properties:
        id:
          type: string
          example: "COMP-213"
        code:
          type: string
          example: "COMP 213"
        name:
          type: string
          example: "Computing for Engineering Technology"
        credits:
          type: number
          example: 3
        semester:
          type: string
          example: "Fall 2025"
        instructor:
          type: object
        description:
          type: string
        schedule:
          type: object
        learningOutcomes:
          type: array
        assessment:
          type: object
        modules:
          type: array
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
paths:
...
  /api/data/courses:
    get:
      operationId: listCourses
      summary: List all courses
      description: Retrieve all courses from the system, optionally filtered by course ID
      tags: [courses]
      parameters:
        - name: id
          in: query
          description: Optional course ID to retrieve a specific course
          schema:
            type: string
            example: "COMP 213"
      responses:
        "200":
          description: List of courses retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        "404":
          description: Course not found (when filtering by id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createCourse
      summary: Create a new course
      description: Add a new course to the system with all course details
      tags: [courses]
      requestBody:
        required: true
        description: Course object containing all course information
        content:
          application/json:
            schema:
              type: object
              required:
                - Course ID
                - Course Name
              properties:
                Course ID:
                  type: string
                  description: Unique course identifier (e.g., "COMP 213")
                Course Name:
                  type: string
                  description: Full course name
                Course Average Weighted Grade:
                  type: string
                  description: Current weighted average grade (optional)
                Credit Hours:
                  type: string
                  description: Number of credit hours
                Course Description:
                  type: string
                  description: Detailed course description
                Professor/Teacher Name:
                  type: string
                  description: Instructor's full name
                Professor/Teacher Email:
                  type: string
                  format: email
                  description: Instructor's email address
              additionalProperties: true
      responses:
        "201":
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        "400":
          description: Bad request - course already exists or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updateCourse
      summary: Update a course
      description: Update an existing course's information. Must include Course ID to identify which course to update.
      tags: [courses]
      requestBody:
        required: true
        description: Complete course object with updated fields
        content:
          application/json:
            schema:
              type: object
              required:
                - Course ID
              properties:
                Course ID:
                  type: string
                  description: Unique course identifier to identify which course to update
                Course Name:
                  type: string
                  description: Full course name
                Course Average Weighted Grade:
                  type: string
                  description: Current weighted average grade
                Credit Hours:
                  type: string
                  description: Number of credit hours
                Course Description:
                  type: string
                  description: Detailed course description
                Professor/Teacher Name:
                  type: string
                  description: Instructor's full name
                Professor/Teacher Email:
                  type: string
                  format: email
                  description: Instructor's email address
              additionalProperties: true
      responses:
        "200":
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        "404":
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteCourse
      summary: Delete a course
      description: Remove a course from the system by its Course ID
      tags: [courses]
      parameters:
        - name: id
          in: query
          required: true
          description: Course ID to delete
          schema:
            type: string
            example: "COMP 213"
      responses:
        "200":
          description: Course deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Detailed Courses API
  /api/courses:
    get:
      operationId: getDetailedCourses
      summary: Get detailed course information
      description: Retrieve all detailed course JSON files or a specific course by ID
      tags: [courses]
      parameters:
        - name: id
          in: query
          description: Optional course ID to retrieve a specific course
          schema:
            type: string
            example: "COMP-213"
      responses:
        "200":
          description: Course or courses retrieved successfully
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/DetailedCourse'
                  - $ref: '#/components/schemas/DetailedCourse'
        "404":
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: createDetailedCourse
      summary: Create a new detailed course
      description: Create a new detailed course JSON file with full course information including modules, learning outcomes, and assessment details
      tags: [courses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetailedCourse'
      responses:
        "201":
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedCourse'
        "400":
          description: Course already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updateDetailedCourse
      summary: Update a detailed course
      description: Update an existing detailed course JSON file with new information
      tags: [courses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetailedCourse'
      responses:
        "200":
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedCourse'
        "404":
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteDetailedCourse
      summary: Delete a detailed course
      description: Delete a course JSON file by course ID
      tags: [courses]
      parameters:
        - name: id
          in: query
          required: true
          description: Course ID to delete
          schema:
            type: string
            example: "COMP-213"
      responses:
        "200":
          description: Course deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "400":
          description: Course ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Quick Links API
...